FW: merge the TestrunSelector's testid_to_test() with sandstone.cpp
This includes a fix to the test list handler not running the test
preinits and group inits.

[ChangeLog][Framework] Fixed a bug that caused certain tests specified
in a test list to not get properly skipped under certain conditions.
